Task5

+++ Rookies/Task-5/Apple_Division.cpp
@@ -3,15 +3,15 @@
 #include <cmath>
 using namespace std;
 
-
-int appleDiv(vector<int> &arr, int endPt, int currentTotal, int total) {
-    if(endPt == 0) {
+long long appleDiv(vector<int> &arr, int endPt, long long currentTotal, long long total) {
+    if (endPt == 0) {
         return abs((total - currentTotal) - currentTotal);
     }
 
-    return min(appleDiv(arr, endPt - 1, currentTotal + arr[endPt - 1], total), 
-            appleDiv(arr, endPt - 1, currentTotal, total));
-
+    return min(
+        appleDiv(arr, endPt - 1, currentTotal + arr[endPt - 1], total),
+        appleDiv(arr, endPt - 1, currentTotal, total)
+    );
 }
 
 int main() {
@@ -19,8 +19,9 @@ int main() {
     cin >> n;
 
     vector<int> arr(n);
-    int total = 0;
-    for(int i = 0; i < n; i++) {
+    long long total = 0;
+
+    for (int i = 0; i < n; i++) {
         cin >> arr[i];
         total += arr[i];
     }
